20kdm2: with two volumes

+++ holly/ta_vertex_parameter.hpp
@@ -684,11 +684,11 @@ namespace ta_vertex_parameter {
     float z;
     float u_0;
     float v_0;
-    uint32_t base_intensity_0;
+    float base_intensity_0;
     float offset_intensity_0;
     float u_1;
     float v_1;
-    uint32_t base_intensity_1;
+    float base_intensity_1;
     float offset_intensity_1;
     uint32_t _res0;
     uint32_t _res1;
@@ -701,11 +701,11 @@ namespace ta_vertex_parameter {
                               const float z,
                               const float u_0,
                               const float v_0,
-                              const uint32_t base_intensity_0,
+                              const float base_intensity_0,
                               const float offset_intensity_0,
                               const float u_1,
                               const float v_1,
-                              const uint32_t base_intensity_1,
+                              const float base_intensity_1,
                               const float offset_intensity_1
                               )
       : parameter_control_word(parameter_control_word)
@@ -756,11 +756,11 @@ namespace ta_vertex_parameter {
     float z;
     uint32_t u_v_0;
     uint32_t _res0;
-    uint32_t base_intensity_0;
+    float base_intensity_0;
     float offset_intensity_0;
     uint32_t u_v_1;
     uint32_t _res1;
-    uint32_t base_intensity_1;
+    float base_intensity_1;
     float offset_intensity_1;
     uint32_t _res2;
     uint32_t _res3;
@@ -772,10 +772,10 @@ namespace ta_vertex_parameter {
                               const float y,
                               const float z,
                               const uint32_t u_v_0,
-                              const uint32_t base_intensity_0,
+                              const float base_intensity_0,
                               const float offset_intensity_0,
                               const uint32_t u_v_1,
-                              const uint32_t base_intensity_1,
+                              const float base_intensity_1,
                               const float offset_intensity_1
                               )
       : parameter_control_word(parameter_control_word)
@@ -1037,4 +1037,3 @@ namespace ta_vertex_parameter {
   static_assert((offsetof (struct modifier_volume, _res5)) == 0x3c);
 
 }
-
